+++ src/werkzeug/sansio/utils.py
@@ -1,7 +1,7 @@
 import typing as t
+from urllib.parse import quote
 
 from .._internal import _encode_idna
-from .._urls import _quote
 from ..exceptions import SecurityError
 from ..urls import uri_to_iri
 
@@ -127,17 +127,19 @@ def get_current_url(
         url.append("/")
         return uri_to_iri("".join(url))
 
-    url.append(_quote(root_path.rstrip("/")))
+    # safe = https://url.spec.whatwg.org/#url-path-segment-string
+    # as well as percent for things that are already quoted
+    url.append(quote(root_path.rstrip("/"), safe="!$&'()*+,/:;=@%"))
     url.append("/")
 
     if path is None:
         return uri_to_iri("".join(url))
 
-    url.append(_quote(path.lstrip("/")))
+    url.append(quote(path.lstrip("/"), safe="!$&'()*+,/:;=@%"))
 
     if query_string:
         url.append("?")
-        url.append(_quote(query_string, safe=":&%="))
+        url.append(quote(query_string, safe="!$&'()*+,/:;=?@%"))
 
     return uri_to_iri("".join(url))